requirements.txt
Flask==3.0.3
gunicorn==23.0.0
flask_sqlalchemy==3.1.1
PyMySQL==1.1.1
alembic==1.13.2
redis==5.0.7
Flask-Admin==1.6.1
WTForms==3.1.2
python-dotenv==1.0.1
aiogram==3.10.0
requests==2.32.3
itsdangerous==2.2.0
PyJWT==2.9.0

